removed unecessary double casting

+++ Project_1/main.cpp
@@ -65,11 +65,11 @@ int main(int numberOfArguments, char *arguments[])
         columnSum = arma::sum(runMatrix, 0);
 
         // Finding averages of trials
-        double time = columnSum(0)/double(trials);
-        double energy = columnSum(1)/double(trials);
-        double energySquared = columnSum(2)/double(trials);
-        double variance = columnSum(3)/double(trials);
-        double acceptanceRatio = columnSum(4)/double(trials);
+        double time = columnSum(0)/trials;
+        double energy = columnSum(1)/trials;
+        double energySquared = columnSum(2)/trials;
+        double variance = columnSum(3)/trials;
+        double acceptanceRatio = columnSum(4)/trials;
 
         std::cout << std::setw(9) << std::setprecision(3) << nParticles
                   << std::setw(12) << std::setprecision(3) << nDimensions
